Struct renames and test deletion
