--- crates/cairo-coverage/src/inner/mod.rs
@@ -1,2 +1,0 @@
-mod config;
-mod data_loader;
