--- crates/cairo-coverage/tests/data/config/empty_versioned_program.json
@@ -1,20 +1,0 @@
-{
-  "version": 1,
-  "type_declarations": [],
-  "libfunc_declarations": [],
-  "statements": [],
-  "funcs": [],
-  "debug_info": {
-    "type_names": [],
-    "libfunc_names": [],
-    "user_func_names": [],
-    "annotations": {
-      "github.com/software-mansion/cairo-profiler": {
-        "statements_functions": {}
-      },
-      "github.com/software-mansion/cairo-coverage": {
-        "statements_code_locations": {}
-      }
-    }
-  }
-}
